% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{slr_rsquared}
\alias{slr_rsquared}
\title{slr_rsquared}
\arguments{
\item{x}{Predictor}

\item{y}{Response}
}
\value{
the r squared of linear regression model for predictor x and response y
}
\description{
Gets the coefficients of the simple linear regression model
}
\examples{
slr_rsquared(c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14),c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69))

}
